+++ src/cqportal/shieldwall/EndCap.py
@@ -1,6 +1,6 @@
 import cadquery as cq
 from cadqueryhelper import Base
-from cqportal.shieldwall import ShieldShape
+from cqportal.shieldwall import ShieldShape, CapGreeble
 
 class EndCap(Base):
     def __init__(self):
@@ -9,13 +9,20 @@ class EndCap(Base):
         self.length = 25
         self.width = 20
         self.height = 25
+        self.base_height = 4
+        self.side_margin = -2
+        self.side_height = 1
+        self.top_height = 2
         
-        self.cut_width = 16
+        self.cut_width = 3
+        self.middle_width_inset = -6
         
-        self.render_greeble = False
+        self.render_greeble = True
+        self.greeble_padding_y = 1
         
         #blueprints
-        self.shape_bp =  ShieldShape()
+        self.shape_bp = ShieldShape()
+        self.greeble_bp = CapGreeble()
         
         #shapes
         self.end_cap = None
@@ -24,6 +31,9 @@ class EndCap(Base):
     def __make_end_cap(self):
         self.shape_bp.length = self.height
         self.shape_bp.width = self.width
+        self.shape_bp.base_height = self.base_height
+        self.shape_bp.middle_width_inset = self.middle_width_inset
+        
         self.shape_bp.make()
         
         self.shape_bp.width = self.height
@@ -36,19 +46,19 @@ class EndCap(Base):
             #.rotate((0,1,0),(0,0,0),90)
         )
         
-        self.shape_bp.width = self.width*2
-        self.shape_bp.base_height += 5
-        self.shape_bp.middle_width_inset = -self.cut_width 
+        self.shape_bp.width = self.length*2
+        self.shape_bp.base_height += self.base_height + self.side_height
+        self.shape_bp.middle_width_inset = -self.length+self.cut_width
         self.shape_bp.length += 4
         self.shape_bp.make()
         cut_shape = (
             self.shape_bp
             .build()
-            .extrude(self.length)
-            .translate((0,0,-1*self.length/2))
+            .extrude(self.width)
+            .translate((0,0,-1*self.width/2))
             .rotate((0,1,0),(0,0,0),90)
             .rotate((0,0,1),(0,0,0),90)
-            .translate((-1*(self.length/2),0,-2))
+            .translate((-1*(self.length/2),0,self.side_margin))
         )
         
         silhouette = (
@@ -61,22 +71,23 @@ class EndCap(Base):
         
         self.end_cap = (
             cq.Workplane('XY')
-            #.union(shape)
             .union(shape.rotate((0,1,0),(0,0,0),90))
-            .cut(silhouette.cut(cut_shape))
+            #.add(cut_shape)
+            .cut(silhouette.cut(cut_shape))#)
         )
         
-        #self.end_cap = silhouette.add(cut_shape)
-        
-    def __make_geeble(self):
-        self.greeble = cq.Workplane("XY").box(10,10,10)
+    def __make_greeble(self):
+        self.greeble_bp.length = self.length - self.cut_width*2 +-1.5
+        self.greeble_bp.width = self.width + self.middle_width_inset*2 - self.greeble_padding_y*2
+        self.greeble_bp.height = self.height - self.base_height - self.side_height - self.top_height
+        self.greeble_bp.make()
         
     def make(self, parent=None):
         super().make()
         self.__make_end_cap()
         
         if self.render_greeble:
-            self.__make_geeble()
+            self.__make_greeble()
         
     def build(self):
         super().build()
@@ -86,5 +97,8 @@ class EndCap(Base):
         )
         
         if self.render_greeble:
-            scene = scene.union(self.greeble)
+            greeble = self.greeble_bp.build()
+            translate_x = self.length/2 - self.greeble_bp.length/2 - self.cut_width
+            translate_z = self.base_height/2+ self.side_height/2 - self.top_height/2
+            scene = scene.add(greeble.translate((-translate_x,0,translate_z)))
         return scene