+++ src/cqportal/shieldwall/__init__.py
@@ -3,4 +3,6 @@ from .StraightBasic import StraightBasic
 from .Mesh import Mesh
 from .Straight import Straight
 from .HexMesh import HexMesh
+from .CapGreeble import CapGreeble
 from .EndCap import EndCap
+from .HexStraight import HexStraight